ugly but almost working good

+++ ex00/BitcoinExchange.hpp
@@ -3,6 +3,8 @@
 #include <fstream>
 #include <map>
 #include <sstream>
+#include <exception>
+#include <typeinfo>
 
 class Exchange
 {	
@@ -11,6 +13,6 @@ class Exchange
 		Exchange(const Exchange &toCopy);
 		~Exchange();
 	private:
-		std::map<std::string, int> values;
+		std::map<std::string, float> values;
 		Exchange &operator=(const Exchange &toCopy);
 };
